Split out into seperate task
